Add StoredCardInputViewControllerTests with UI and interaction test coverage
