+++ Tests/GeneratedMocks/AutoMockable.generated.swift
@@ -17,8 +17,10 @@ import Foundation
 #endif
 
 @testable import Adyen
+@testable import AdyenCard
 @testable import AdyenCheckout
 @testable import AdyenDropIn
+@testable import AdyenUI
 
 class ActionPresenterMock: ActionPresenter {
 
@@ -580,5 +582,91 @@ class RouterMock: Router {
     }
 
     var underlyingRootViewController: UIViewController!
+
+}
+
+class StoredCardInputViewModelProtocolMock: StoredCardInputViewModelProtocol {
+
+    var cardImageItem: CardImageItem {
+        get { underlyingCardImageItem }
+        set(value) { underlyingCardImageItem = value }
+    }
+
+    var underlyingCardImageItem: CardImageItem!
+    var titleText: String {
+        get { underlyingTitleText }
+        set(value) { underlyingTitleText = value }
+    }
+
+    var underlyingTitleText: String!
+    var subtitleText: NSAttributedString {
+        get { underlyingSubtitleText }
+        set(value) { underlyingSubtitleText = value }
+    }
+
+    var underlyingSubtitleText: NSAttributedString!
+    var securityCodeItem: FormCardSecurityCodeItem {
+        get { underlyingSecurityCodeItem }
+        set(value) { underlyingSecurityCodeItem = value }
+    }
+
+    var underlyingSecurityCodeItem: FormCardSecurityCodeItem!
+    var submitButtonTitle: String {
+        get { underlyingSubmitButtonTitle }
+        set(value) { underlyingSubmitButtonTitle = value }
+    }
+
+    var underlyingSubmitButtonTitle: String!
+    var theme: AdyenTheme {
+        get { underlyingTheme }
+        set(value) { underlyingTheme = value }
+    }
+
+    var underlyingTheme: AdyenTheme!
+    var onViewInstruction: Completion<StoredCardInputViewInstruction>?
+
+    // MARK: - submit
+
+    var submitCallsCount = 0
+    var submitCalled: Bool {
+        submitCallsCount > 0
+    }
+
+    var submitClosure: (() async -> Void)?
+
+    @MainActor
+    func submit() async {
+        submitCallsCount += 1
+        await submitClosure?()
+    }
+
+    // MARK: - dismiss
+
+    var dismissCallsCount = 0
+    var dismissCalled: Bool {
+        dismissCallsCount > 0
+    }
+
+    var dismissClosure: (() -> Void)?
+
+    @MainActor
+    func dismiss() {
+        dismissCallsCount += 1
+        dismissClosure?()
+    }
+
+    // MARK: - viewDidLoad
+
+    var viewDidLoadCallsCount = 0
+    var viewDidLoadCalled: Bool {
+        viewDidLoadCallsCount > 0
+    }
+
+    var viewDidLoadClosure: (() -> Void)?
+
+    func viewDidLoad() {
+        viewDidLoadCallsCount += 1
+        viewDidLoadClosure?()
+    }
 
 }
