+++ Tests/IntegrationTests/Card Tests/StoredCardInputViewModelTests.swift
@@ -256,345 +256,3 @@ enum StoredCardTestData {
         AmountData(amount: Amount(value: 0, currencyCode: "GBP"), expectedFormatted: "£0.00")
     ]
 }
-
-// MARK: - StoredCardInputViewControllerTests
-
-@MainActor
-struct StoredCardInputViewControllerTests {
-
-    // MARK: - A: UI Display
-
-    @Test
-    func viewDidLoad_callsViewModelViewDidLoad() {
-        let (proxy, viewModel) = makeSUT()
-        proxy.load()
-        #expect(viewModel.viewDidLoadCallsCount == 1)
-    }
-
-    @Test
-    func viewDidLoad_configuresLabelsAndButtonsFromViewModel() throws {
-        let titleText = "Enter security code"
-        let subtitleText = "Use your Visa card"
-        let submitTitle = "Pay €9.99"
-        let otherTitle = "Other payment options"
-
-        let (proxy, _) = makeSUT(
-            titleText: titleText,
-            subtitleText: subtitleText,
-            submitButtonTitle: submitTitle,
-            showAllPaymentMethodsButtonTitle: otherTitle
-        )
-        proxy.load()
-
-        #expect(try proxy.titleLabelText == titleText)
-        #expect(try proxy.subtitleLabelText == subtitleText)
-        #expect(try proxy.primaryButtonTitle == submitTitle)
-        #expect(try proxy.secondaryButtonTitle == otherTitle)
-    }
-
-    @Test
-    func viewDidLoad_setsIsModalInPresentation() {
-        let (proxy, _) = makeSUT()
-        proxy.load()
-        #expect(proxy.viewController.isModalInPresentation)
-    }
-
-    @Test
-    func viewDidLoad_hasBackNavigationButton() {
-        let (proxy, _) = makeSUT()
-        proxy.load()
-        #expect(proxy.viewController.navigationItem.leftBarButtonItem != nil)
-    }
-
-    // MARK: - B: Primary button tap
-
-    @Test
-    func primaryButtonTap_callsSubmit() async throws {
-        let (proxy, viewModel) = makeSUT()
-        proxy.load()
-
-        try proxy.tapPrimaryButton()
-
-        // submit is dispatched inside a Task — yield to let it run
-        for _ in 0..<10 {
-            await Task.yield()
-        }
-
-        #expect(viewModel.submitCallsCount == 1)
-    }
-
-    // MARK: - C: View instructions → UI
-
-    @Test
-    func viewInstruction_setLoadingTrue_disablesButtonAndShowsSpinner() throws {
-        let (proxy, viewModel) = makeSUT()
-        proxy.load()
-
-        viewModel.onViewInstruction?(.setLoading(true))
-
-        let button = try proxy.primaryButton()
-        #expect(button.showsActivityIndicator)
-        #expect(!button.isEnabled)
-    }
-
-    @Test
-    func viewInstruction_setLoadingFalse_enablesButtonAndHidesSpinner() throws {
-        let (proxy, viewModel) = makeSUT()
-        proxy.load()
-
-        viewModel.onViewInstruction?(.setLoading(true))
-        viewModel.onViewInstruction?(.setLoading(false))
-
-        let button = try proxy.primaryButton()
-        #expect(!button.showsActivityIndicator)
-        #expect(button.isEnabled)
-    }
-
-    @Test
-    func viewInstruction_showSecurityCodeValidation_setsInvalidStateOnItem() {
-        let (proxy, viewModel) = makeSUT()
-        proxy.load()
-
-        viewModel.onViewInstruction?(.showSecurityCodeValidation)
-
-        if case .invalid = viewModel.securityCodeItem.validationState {
-            // pass
-        } else {
-            Issue.record("Expected validationState to be .invalid after showSecurityCodeValidation instruction")
-        }
-    }
-
-    // MARK: - D: Secondary button and back button
-
-    @Test
-    func secondaryButtonTap_callsShowAllPaymentMethods() throws {
-        let (proxy, viewModel) = makeSUT()
-        proxy.load()
-
-        try proxy.tapSecondaryButton()
-
-        #expect(viewModel.showAllPaymentMethodsCallsCount == 1)
-    }
-
-    @Test
-    func backButtonTap_callsDismiss() {
-        let (proxy, viewModel) = makeSUT()
-        proxy.load()
-
-        proxy.tapBackButton()
-
-        #expect(viewModel.dismissCallsCount == 1)
-    }
-
-    // MARK: - Helpers
-
-    private func makeSUT(
-        titleText: String = "Enter security code",
-        subtitleText: String = "Use your Visa card",
-        submitButtonTitle: String = "Pay €1.00",
-        showAllPaymentMethodsButtonTitle: String = "Other payment options"
-    ) -> (proxy: StoredCardInputViewControllerProxy, viewModel: StoredCardInputViewModelProtocolMock) {
-        let viewModel = StoredCardInputViewModelProtocolMock()
-        viewModel.underlyingTitleText = titleText
-        viewModel.underlyingSubtitleText = NSAttributedString(string: subtitleText)
-        viewModel.underlyingSubmitButtonTitle = submitButtonTitle
-        viewModel.underlyingShowAllPaymentMethodsButtonTitle = showAllPaymentMethodsButtonTitle
-        viewModel.underlyingTheme = .default
-        viewModel.underlyingSecurityCodeItem = FormCardSecurityCodeItem()
-        viewModel.underlyingCardImageItem = CardImageItem(
-            imageURL: nil,
-            sizeMode: .fixed(CGSize(width: 80, height: 52)),
-            theme: .default
-        )
-
-        let viewController = StoredCardInputViewController(viewModel: viewModel)
-        return (StoredCardInputViewControllerProxy(viewController: viewController), viewModel)
-    }
-}
-
-// MARK: - StoredCardInputViewControllerProxy
-
-@MainActor
-struct StoredCardInputViewControllerProxy {
-    let viewController: StoredCardInputViewController
-
-    func load() {
-        viewController.loadViewIfNeeded()
-    }
-
-    var titleLabelText: String {
-        get throws {
-            let label = try #require(
-                viewController.view.findView(by: "title") as? UILabel,
-                "Cannot find title label"
-            )
-            return try #require(label.text)
-        }
-    }
-
-    var subtitleLabelText: String {
-        get throws {
-            let label = try #require(
-                viewController.view.findView(by: "subTitle") as? UILabel,
-                "Cannot find subtitle label"
-            )
-            return try #require(label.attributedText?.string)
-        }
-    }
-
-    var primaryButtonTitle: String {
-        get throws {
-            let button = try primaryButton()
-            return try #require(button.title)
-        }
-    }
-
-    var secondaryButtonTitle: String {
-        get throws {
-            let button = try secondaryButton()
-            return try #require(button.title)
-        }
-    }
-
-    func primaryButton() throws -> FormButton {
-        try #require(
-            viewController.view.findView(by: "primaryButton") as? FormButton,
-            "Cannot find primaryButton"
-        )
-    }
-
-    func secondaryButton() throws -> FormButton {
-        try #require(
-            viewController.view.findView(by: "secondaryButton") as? FormButton,
-            "Cannot find secondaryButton"
-        )
-    }
-
-    func tapPrimaryButton() throws {
-        try primaryButton().sendActions(for: .touchUpInside)
-    }
-
-    func tapSecondaryButton() throws {
-        try secondaryButton().sendActions(for: .touchUpInside)
-    }
-
-    func tapBackButton() {
-        let backButton = viewController.navigationItem.leftBarButtonItem
-        _ = backButton?.target?.perform(backButton?.action)
-    }
-}
-
-// MARK: - StoredCardInputViewModelProtocolMock
-
-class StoredCardInputViewModelProtocolMock: StoredCardInputViewModelProtocol {
-
-    var cardImageItem: CardImageItem {
-        get { underlyingCardImageItem }
-        set(value) { underlyingCardImageItem = value }
-    }
-
-    var underlyingCardImageItem: CardImageItem!
-
-    var titleText: String {
-        get { underlyingTitleText }
-        set(value) { underlyingTitleText = value }
-    }
-
-    var underlyingTitleText: String! = ""
-
-    var subtitleText: NSAttributedString {
-        get { underlyingSubtitleText }
-        set(value) { underlyingSubtitleText = value }
-    }
-
-    var underlyingSubtitleText: NSAttributedString! = NSAttributedString()
-
-    var securityCodeItem: FormCardSecurityCodeItem {
-        get { underlyingSecurityCodeItem }
-        set(value) { underlyingSecurityCodeItem = value }
-    }
-
-    var underlyingSecurityCodeItem: FormCardSecurityCodeItem! = FormCardSecurityCodeItem()
-
-    var submitButtonTitle: String {
-        get { underlyingSubmitButtonTitle }
-        set(value) { underlyingSubmitButtonTitle = value }
-    }
-
-    var underlyingSubmitButtonTitle: String! = ""
-
-    var showAllPaymentMethodsButtonTitle: String {
-        get { underlyingShowAllPaymentMethodsButtonTitle }
-        set(value) { underlyingShowAllPaymentMethodsButtonTitle = value }
-    }
-
-    var underlyingShowAllPaymentMethodsButtonTitle: String! = ""
-
-    var theme: AdyenTheme {
-        get { underlyingTheme }
-        set(value) { underlyingTheme = value }
-    }
-
-    var underlyingTheme: AdyenTheme! = .default
-
-    var onViewInstruction: ((StoredCardInputViewInstruction) -> Void)?
-
-    // MARK: - submit
-
-    var submitCallsCount = 0
-    var submitCalled: Bool {
-        submitCallsCount > 0
-    }
-
-    var submitClosure: (() async -> Void)?
-
-    @MainActor
-    func submit() async {
-        submitCallsCount += 1
-        await submitClosure?()
-    }
-
-    // MARK: - showAllPaymentMethods
-
-    var showAllPaymentMethodsCallsCount = 0
-    var showAllPaymentMethodsCalled: Bool {
-        showAllPaymentMethodsCallsCount > 0
-    }
-
-    var showAllPaymentMethodsClosure: (() -> Void)?
-
-    @MainActor
-    func showAllPaymentMethods() {
-        showAllPaymentMethodsCallsCount += 1
-        showAllPaymentMethodsClosure?()
-    }
-
-    // MARK: - dismiss
-
-    var dismissCallsCount = 0
-    var dismissCalled: Bool {
-        dismissCallsCount > 0
-    }
-
-    var dismissClosure: (() -> Void)?
-
-    @MainActor
-    func dismiss() {
-        dismissCallsCount += 1
-        dismissClosure?()
-    }
-
-    // MARK: - viewDidLoad
-
-    var viewDidLoadCallsCount = 0
-    var viewDidLoadCalled: Bool {
-        viewDidLoadCallsCount > 0
-    }
-
-    var viewDidLoadClosure: (() -> Void)?
-
-    func viewDidLoad() {
-        viewDidLoadCallsCount += 1
-        viewDidLoadClosure?()
-    }
-}
